EntryPointHaxx v1
The reason behind me releasing this source is because of the developer completely leaving ROBLOX as a whole, and every single obfuscator I tried literally broke the script.
I'm releasing this source for anyone to new user off of or do anything, please do not sell it though. Most of the scripts are made by other users so don't call me a skid. I made this script for fun, and will continue to update this script in this specific file. Thanks!